requirements.txt
pandas
matplotlib
seaborn
folium
jupyter
ipykernel<7
requests
boto3
pyarrow
meteostat
pyspark==3.5.6
numpy
scikit-learn
streamlit
streamlit-folium
plotly
kafka-python
